refactor: user update form
